pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-txt-generator"
version = "0.1.0"
description = "URL to llm.txt generator - crawl docs sites and create LLM-friendly summaries"
authors = [{name = "LLM-TXT Team"}]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "html2text>=2020.1.16",
    "markdownify>=0.11.0",
    "cohere>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "playwright>=1.40.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "mcp>=0.1.0",
    "aioboto3>=12.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "types-beautifulsoup4>=4.12.0"
]

[tool.ruff]
target-version = "py39"
line-length = 88
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PGH", "PL", "TRY", "RUF"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["llm_txt"]

[project.scripts]
llm-txt = "llm_txt.cli:main"
llmxt = "llm_txt.cli.llmxt:main"
llm-txt-mcp = "llm_txt.mcp.server:main"
